bej_c1_s1_maven_pc_1.2/B_C1_S1_Maven_Pc_1.2/src/main/java/org/niit/App.java
package org.niit;

public class App 
{
    public int sum(int number1,int number2)
    {
        return number1+number2;
    }

    public int minus(int number1,int number2)
    {
        return number1-number2;
    }

    public int multiply(int number1,int number2)
    {
        return number1*number2;
    }

    public int divide(int number1,int number2)
    {
        if(number2==0)
            throw new IllegalArgumentException("can not divided by zero");
        return number1/number2;
    }
}
